Refactoring some code and add config reference link to locales file
